Move Version() to Cluster; add proper tests
